moved map analyses to service
